// src/app/models/api.ts
export interface ProductDto {
  id?: number;
  description?: string | null;
  barcode?: string | null;
  companyId?: number;
  cost: number;
  stock: number;
  shortDescription?: string | null;
  reference?: string | null;
  maximumQuantity?: number;
  minimumQuantity?: number;
  productTypeId?: number;
  categoryId?: number;
  unitOfMeasurementId?: number;
  invoiceWithoutStock?: boolean;
}
// ...
export interface SaleResponseDto {
  id: number;
  customerId?: number | null;
  total: number;
  paidAmount: number;
  cashRegisterId?: number | null;
  dueDate?: string | null;
  isCancelled: boolean;
  creationDate: string;
  details: {
    productId: number;
    quantity: number;
    unitPrice: number;
  }[];
}